qjazz-processes/pyproject.toml
[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[project]
name = "qjazz-processes"
description = "QJazz OGC processes "
requires-python = ">= 3.12"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX",
]
dependencies = [
    "qjazz-contrib[celery]==0.4.0",
    "pydantic-extra-types",
    "click",
    "requests",
    "aiofiles",
    "ruamel.yaml",
]
version = "0.4.0"

[project.optional-dependencies]
server = ["aiohttp"]
minio = ["minio<7.2.16"]
mail = ["email-validator"]

[[project.authors]]
name = "David Marteau"
email = "dmarteau@3liz.org"

[project.scripts]
qjazz-processes = "qjazz_processes.__main__:main"
qjazz-processing = "qjazz_processing.__main__:main"

[tool.uv.build-backend]
module-name = [ "qjazz_processes", "qjazz_processing"]

[tool.uv.sources]
qjazz-contrib = { workspace = true }

#
# Ruff
#

[tool.ruff]
extend = "../config/ruff.toml"
extend-exclude = [
    "tests/plugins/",
    "tests/clienttests/",
] 

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T201"]
"src/qjazz_processes/tests/*" = ["T201"]
"src/qjazz_processes/server/_handlers/protos.py" = ["F401"]
"src/qjazz_processes/processing/prelude.py" = ["F401"]

